requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.31.0
Flask-WTF==...

psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1

pytest==7.4.3